+++ src/Manager.php
@@ -32,6 +32,21 @@ class Manager implements
     use ManagesItemsTrait, ChainsNestedItemsTrait;
 
     /**
+     * The items stored in the manager
+     * @var array $items Items governed by manager
+     */
+    protected $items;
+
+    /**
+     * Build a new manager instance
+     * @param array $items
+     */
+    public function __construct($items = [])
+    {
+        $this->initManager($items);
+    }
+
+    /**
      * @implements ArrayAccess
      */
     public function offsetExists($offset)