+++ tests/Stubs/ChainsNestedItemsTraitStub.php
@@ -12,4 +12,18 @@ class ChainsNestedItemsTraitStub
 {
     use ManagesItemsTrait, ChainsNestedItemsTrait;
 
+    /**
+     * The items stored in the manager
+     * @var array $items Items governed by manager
+     */
+    protected $items;
+
+    /**
+     * Build a new manager instance
+     * @param array $items
+     */
+    public function __construct($items = [])
+    {
+        $this->initManager($items);
+    }
 }